opsoro_SERVER3/opsoro/requirements.txt
django-user-accounts
social-auth-app-django
pinax-webanalytics
pinax-eventlog

Django
pinax-theme-bootstrap

pinax-announcements
django-mysql
django-rosetta

django-cors-headers
Pillow
psycopg2
stripe

channels
asgi_redis
